Volunteers done and made

+++ app/Http/Controllers/VolunteerController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\File;
 use App\Models\Volunteer;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -15,7 +16,12 @@ class VolunteerController extends Controller
      */
     public function index()
     {
-        //
+        $volunteers = Volunteer::orderBy('name', 'asc')->simplePaginate(15);
+        $volunteerImages = [];
+        foreach ($volunteers as $volunteer) {
+            array_push($volunteerImages, Storage::url($volunteer->path));
+        }
+        return view('pages.dashboard.admin.designs.volunteers.indexVolunteers', ['volunteers' => $volunteers, 'volunteerImages' => $volunteerImages]);
     }
 
     /**
@@ -25,7 +31,7 @@ class VolunteerController extends Controller
      */
     public function create()
     {
-        //
+        return view('pages.dashboard.admin.designs.volunteers.createVolunteer');
     }
 
     /**
@@ -44,8 +50,12 @@ class VolunteerController extends Controller
             'top' => 'required|boolean',
             'information' => 'required|max:300'
         ]);
-        $link = $valid['file']->store('/public');
-        $urlLocal = Storage::url($link);
+        //save file
+        $link = null;
+        if ($request->hasFile('file')) {
+            $link = $valid['file']->store('/public/volunteers');
+        }
+
         $volunteer = new Volunteer();
         $volunteer->name = $valid['name'];
         $volunteer->email = $valid['email'];
@@ -53,9 +63,8 @@ class VolunteerController extends Controller
         $volunteer->top = $valid['top'];
         $volunteer->information = $valid['information'];
         $volunteer->path = $link;
-        $volunteer->localUrl = $urlLocal;
         $volunteer->save();
-        return redirect("/cms/volunteers/create")->with('success', "De vrijwiliger is succesvol aangemaakt");
+        return back()->with('success', "De vrijwiliger is succesvol aangemaakt");
     }
 
     /**
@@ -64,9 +73,14 @@ class VolunteerController extends Controller
      * @param  \App\Models\Volunteer  $volunteer
      * @return \Illuminate\Http\Response
      */
-    public function show(Volunteer $volunteer)
+    public function show($id)
     {
-        //
+        if ($id === null) {
+            return redirect('/cms/volunteers')->with('error', "Oeps er ging iets miss");
+        }
+        $volunteer = Volunteer::find($id);
+        $image = Storage::url($volunteer->path);
+        return view('pages.dashboard.admin.designs.volunteers.indexVolunteer', ['volunteer' => $volunteer, 'image' => $image]);
     }
 
     /**
@@ -75,9 +89,14 @@ class VolunteerController extends Controller
      * @param  \App\Models\Volunteer  $volunteer
      * @return \Illuminate\Http\Response
      */
-    public function edit(Volunteer $volunteer)
+    public function edit($id)
     {
-        //
+        if ($id === null) {
+            return redirect('/cms/volunteers')->with('error', "Oeps er ging iets miss");
+        }
+        $volunteer = Volunteer::find($id);
+        $image = Storage::url($volunteer->path);
+        return view('pages.dashboard.admin.designs.volunteers.editVolunteer', ['volunteer' => $volunteer, 'image' => $image]);
     }
 
     /**
@@ -87,9 +106,40 @@ class VolunteerController extends Controller
      * @param  \App\Models\Volunteer  $volunteer
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Volunteer $volunteer)
+    public function update(Request $request, $id)
     {
-        //
+        if ($id === null) {
+            return redirect('/cms/volunteers')->with('error', "Oeps er ging iets miss");
+        }
+        $volunteer = Volunteer::find($id);
+
+        $valid = $request->validate([
+            'name' => 'required|string|unique:volunteers,name,' . $volunteer->id,
+            'email' => 'email|required',
+            'phoneNumber' => 'required',
+            'file' => 'max:4096',
+            'top' => 'required|boolean',
+            'information' => 'required|max:300'
+        ]);
+
+        if ($request->hasFile('file')) {
+            Storage::disk('local')->delete($volunteer->path);
+        }
+
+        $volunteer->name = $valid['name'];
+        $volunteer->email = $valid['email'];
+        $volunteer->phoneNumber = $valid['phoneNumber'];
+        $volunteer->top = $valid['top'];
+        $volunteer->information = $valid['information'];
+        $volunteer->save();
+
+        if ($request->hasFile('file')) {
+            $link = $valid['file']->store('/public/volunteers');
+            $volunteer->path = $link;
+        }
+
+        $volunteer->save();
+        return back()->with('success', "De wijziging was successvol");
     }
 
     /**
@@ -104,7 +154,9 @@ class VolunteerController extends Controller
             return back()->with('error', "Oeps er ging iets fout");
         }
         $volunteer = Volunteer::find($id);
-        Storage::disk('local')->delete($volunteer->path);
+        if ($volunteer->path !== null) {
+            Storage::disk('local')->delete($volunteer->path);
+        }
         $message = "U heeft succesvol de vrijwiliger $volunteer->name verwijdert";
         $volunteer->delete();
         return back()->with('success', $message);
